# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ZIAY API
  version: 1.0.0
  description: |
    Comercio Conversacional + Atribución Inteligente.

    Multi-tenant conversational-commerce platform with agent-driven checkout
    (AP2 / UCP / ACP / MCP protocol trinity), ad attribution, compliance
    (Ley 1581 / Ley 2573 / Ley 1480 / DIAN), and a wallet/commission layer
    for traffickers + marketplaces.

    All tenant-scoped endpoints require a session cookie (`sessionAuth`)
    unless marked otherwise. UCP / ACP / webhook endpoints use `bearerAuth`
    or are public.

  contact:
    email: api@ziay.co
  license:
    name: Proprietary
servers:
  - url: https://ziay.co
    description: Production
  - url: http://localhost:3000
    description: Development
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth v4 session cookie (set by /api/auth/[...nextauth]/callback/credentials).
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed W3C Verifiable Credential or bearer token (UCP/AP2/ACP).
    cronAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Bearer $CRON_SECRET — shared secret for /api/compliance/retention/cron.'
    webhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: HMAC-SHA256 signature from the gateway (Meta/Stripe/Wompi/etc.).
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
...
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - sessionAuth: []
paths:
...
  /api/ai-reply:
    post:
      summary: Generate AI draft reply
      description: Generate an AI draft reply for a conversation message (uses the tenant's LLM adapter).
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: AI-generated draft reply.

  # ─────────────────────────────────────────────────────────────────────
  # 6. CHANNELS & INTEGRATIONS
  # ─────────────────────────────────────────────────────────────────────
...
  /api/ap2/mandates/{id}:
    get:
      summary: Fetch mandate by id
      description: Fetch a single AP2 mandate (full VC payload + signature verification status).
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mandate object including signed VC.
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update mandate
      description: Advance status (active→consumed), stamp orderId / paymentRef.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, consumed]
                orderId:
                  type: string
                paymentRef:
                  type: string
      responses:
        '200':
          description: Updated mandate.
  /api/ap2/mandates/{id}/revoke:
    patch:
      summary: Revoke mandate (with cascade)
      description: Revoke an AP2 mandate. Cascades to children (Intent → Cart → Payment) in a single transaction.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: '{ mandateId, status: "revoked", revokedAt, reason }'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Already revoked.
...